feat: move sink rewriter to lib.rs
